Starting to implement like button functionality

+++ blog/api/controllers/blogs.js
@@ -1,21 +1,28 @@
 const blogsRouter = require('express').Router()
+const blog = require('../models/blog')
 const Blog = require('../models/blog')
+const { userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (req, res) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   res.status(200).json(blogs)
 })
 
-blogsRouter.post('/', async (req, res) => {
+blogsRouter.post('/', userExtractor, async (req, res) => {
+  const { user } = req
   const { title, author, url, likes } = req.body
 
   const blog = new Blog({
     title: title,
     author: author,
     url: url,
+    user: user,
     likes: likes,
   })
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save({ validateModifiedOnly: true })
+
   res.status(201).json(savedBlog)
 })
 
@@ -32,9 +39,19 @@ blogsRouter.put('/:id', async (req, res) => {
   res.status(201).json(updatedBlog)
 })
 
-blogsRouter.delete('/:id', async (req, res) => {
-  await Blog.findByIdAndDelete(req.params.id)
-  res.status(204).end()
+blogsRouter.delete('/:id', userExtractor, async (req, res) => {
+  const { user } = req
+
+  if (!user) {
+    return res.status(401).json({ error: 'token invalid' })
+  }
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return res.status(401).json({ error: 'unauthorized access' })
+  } else {
+    await Blog.findByIdAndDelete(req.params.id)
+    res.status(204).end()
+  }
 })
 
 module.exports = blogsRouter
